added test for ingredients selector
