+++ package.json
@@ -30,16 +30,13 @@
     "zod": "^3.23.8"
   },
   "devDependencies": {
-    "@happy-dom/global-registrator": "^15.11.6",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.0.1",
-    "@testing-library/user-event": "^14.5.2",
     "@types/bun": "^1.1.13",
     "@types/jest": "^29.5.14",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
-    "@types/testing-library__jest-dom": "^6.0.0",
     "eslint": "^8",
     "eslint-config-next": "15.0.3",
     "jest-environment-jsdom": "^29.7.0",
