--- tests/setup.ts
@@ -1,4 +1,0 @@
-import { expect, test } from "bun:test";
-import * as matchers from '@testing-library/jest-dom/matchers';
-
-expect.extend(matchers);
