--- tests/testSetup.ts
@@ -1,2 +1,0 @@
-// testSetup.ts
-import "@testing-library/jest-dom";
